test now shows bug and is closer to minimal
