+++ test/test_sparql_extended/test_unbound_values-working.rq
@@ -1,0 +1,10 @@
+PREFIX ex: <http://example.org/>
+PREFIX foaf: <http://xmlns.com/foaf/0.1/>
+
+SELECT ?person (MAX(DISTINCT ?age) AS ?maxAge)
+WHERE {
+    ?person a ex:Person .
+
+    OPTIONAL { ?person foaf:age ?age . }
+}
+GROUP BY ?person
